打开相机或相册Activity封装，BaseActivity修改
