+++ app/src/main/java/com/wentongwang/mysports/constant/IntentConstants.java
@@ -14,4 +14,5 @@ public class IntentConstants {
     //permission
     public static final int MY_PERMISSIONS_REQUEST_TAKE_PHOTO = 1;
     public static final int MY_PERMISSIONS_REQUEST_CHOOSE_PHOTO = 2;
+    public static final int MY_PERMISSIONS_REQUEST_CHOOSE_MULTI_PHOTO = 3;
 }
