+++ app/src/main/java/com/wentongwang/mysports/views/activity/createvent/CreatEventActivity.java
@@ -61,7 +61,6 @@ public class CreatEventActivity extends BaseActivity implements CreatEventView {
         listGViews = new ArrayList<>();
 
         initSportEvents();
-        ButterKnife.bind(CreatEventActivity.this);
         initGirdViewsAndPoints();
         mGridViewContainer.setAdapter(new MyViewPagerAdapter());
